requirements.txt
Django==5.1.2
django-extensions==3.2.3
celery==5.4.0
psycopg2_binary==2.9.10
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
